Refactor k-dropzone
