+++ panel/src/config/components.js
@@ -206,7 +206,7 @@ import Card from "@/components/Layout/Card.vue";
 import Cards from "@/components/Layout/Cards.vue";
 import Collection from "@/components/Layout/Collection.vue";
 import Column from "@/components/Layout/Column.vue";
-import Dropzone from "@/components/Layout/Dropzone.vue";
+import Dropzone from "@/ui/components/Dropzone.vue";
 import Empty from "@/components/Layout/Empty.vue";
 import FilePreview from "@/components/Layout/FilePreview.vue";
 import Grid from "@/components/Layout/Grid.vue";
