+++ panel/src/ui/components/Dropzone.vue
@@ -2,7 +2,7 @@
   <div
     :data-dragging="dragging"
     :data-over="over"
-    class="k-dropzone"
+    class="k-dropzone relative"
     @dragenter="onEnter"
     @dragleave="onLeave"
     @dragover="onOver"
@@ -15,10 +15,9 @@
 <script>
 export default {
   props: {
-    label: {
-      type: String,
-      default: "Drop to upload"
-    },
+    /**
+     * You can deactivate the dropzone with this prop
+     */
     disabled: {
       type: Boolean,
       default: false
@@ -48,9 +47,8 @@ export default {
         return this.reset();
       }
 
-      this.$events.$emit("dropzone.drop");
-
       this.files = $event.dataTransfer.files;
+      this.$events.$emit("dropzone.drop");
       this.$emit("drop", this.files);
       this.reset();
     },
@@ -73,9 +71,6 @@ export default {
 </script>
 
 <style lang="scss">
-.k-dropzone {
-  position: relative;
-}
 .k-dropzone::after {
   content: "";
   position: absolute;
@@ -92,5 +87,4 @@ export default {
   outline: 1px solid $color-focus;
   box-shadow: $color-focus-outline 0 0 0 3px;
 }
-
 </style>
